requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
Werkzeug==2.0.2
SQLAlchemy==1.3.23